Move print of statistic to separate class
