+++ src/CMakeLists.txt
@@ -35,6 +35,7 @@ add_library(crossing_consoles_lib
         communication/message_layer/message/PayloadMessage.cpp
         communication/message_layer/message/MessageMetaData.cpp
         communication/message_layer/message/AcknowledgeMessage.cpp
+        communication/message_layer/message/ConnectionResetMessage.cpp
 
         communication/message_layer/crc/CRCHandler.cpp
 
@@ -42,13 +43,15 @@ add_library(crossing_consoles_lib
         communication/connection_layer/connection/ConnectionManager.cpp
         communication/connection_layer/connection/ServerSideConnectionManager.cpp
         communication/connection_layer/connection/ClientSideConnectionManager.cpp
-        communication/connection_layer/connection/ConnectionStatistics.cpp
+
+        communication/connection_layer/connection/statistics/ConnectionStatistics.cpp
+        communication/connection_layer/connection/statistics/StatisticPrinter.cpp
 
         communication/connection_layer/event/Event.cpp
         communication/connection_layer/event/ConnectEvent.cpp
         communication/connection_layer/event/DisconnectEvent.cpp
         communication/connection_layer/event/PayloadEvent.cpp
-        communication/message_layer/message/ConnectionResetMessage.cpp
+
         )
 
 add_executable(crossing_consoles_main_server
