+++ src/communication/connection_layer/connection/statistics/ConnectionStatistics.h
@@ -4,7 +4,7 @@
 #include <list>
 #include <unordered_map>
 
-#include "../../message_layer/message/Message.h"
+#include "../../../message_layer/message/Message.h"
 
 namespace communication::connection_layer {
 
@@ -16,6 +16,20 @@ class ConnectionStatistics {
 
  public:
   ConnectionStatistics();
+
+  // Variables used for the MessageType statistics and the calculation of the package loss
+  typedef uint16_t message_count_t;
+  typedef std::unordered_map<message_layer::MessageType, message_count_t> message_count_map_t;
+
+  struct MessageStatisticData {
+    message_count_map_t map = {};
+    message_count_t count = 0;
+  };
+
+  struct PackageLossData {
+    message_count_t package_loss = 0;
+    double package_loss_percentage = 0;
+  };
 
   /**
    * \brief Add received message to statistics.
@@ -33,58 +47,24 @@ class ConnectionStatistics {
   void SetReceivedTimestampForSentMessage(ProtocolDefinition::sequence_t sequence);
 
   /**
-   * \brief Print statistics to console.
-   * \details Prints message count (sent, received) for each message type plus package loss, average response time and
-   * uptime of the Connection.
-   */
-  void PrintStatistics();
-
- private:
-  /**
-   * \brief Print message count for each Message Type to console.
-   */
-  void PrintMapStatistics();
-
-  /**
    * \brief Calculate overall package loss.
    */
-  void CalculatePackageLoss();
-
-  /**
-   * \brief Print overall package loss to console.
-   */
-  void PrintPackageLoss() const;
+  PackageLossData CalculatePackageLoss();
 
   /**
    * \brief Calculate the average response time for all Messages in milliseconds
    */
   double CalculateAverageResponseTime();
-
-  /**
-   * \brief Print the average response time in milliseconds
-   */
-  void PrintAverageResponseTime();
-
   /**
    * \brief Calculate the uptime of the Connection in milliseconds
    */
   double CalculateUptime();
 
-  /**
-   * \brief Print the uptime of the Connection in milliseconds
-   */
-  void PrintUptime();
+  MessageStatisticData GetReceivedMessageStatistics();
 
-  // Variables used for the MessageType statistics and the calculation of the package loss
-  typedef uint16_t message_count_t;
-  typedef std::unordered_map<message_layer::MessageType, message_count_t> message_count_map_t;
+  MessageStatisticData GetSentMessageStatistics();
 
-  /// Struct to hold the total of all messages and for each message type
-  struct MessageStatisticData {
-    message_count_map_t map = {};
-    message_count_t count = 0;
-  };
-
+ private:
   /// Hold the total of all received messages and for each message type
   MessageStatisticData received_message_statistics;
 
@@ -96,17 +76,11 @@ class ConnectionStatistics {
    */
   static void UpdateStatisticData(MessageStatisticData& message_statistics, const message_layer::Message& message);
 
-  message_count_t package_loss = 0;
-  double package_loss_percentage = 0;
-
   /// List of all sent messages which is used to calculate the average response time
   std::list<message_layer::Message> sent_message_list;
 
   /// Time when the Connection and the Connection statistic class were created, used for the calculation of the uptime
   std::chrono::steady_clock::time_point connection_start_time;
-
-  /// Mapping from MessageTypes to Strings, used for printing of the statistics
-  static const std::unordered_map<message_layer::MessageType, const std::string> map_message_type_to_string;
 };
 
 }  // namespace communication::connection_layer
